Switch to CMake and cleanup
