+++ .gitignore
@@ -1,60 +1,2 @@
-# Windows image file caches
-Thumbs.db
-ehthumbs.db
-
-# Folder config file
-Desktop.ini
-
-# Recycle Bin used on file shares
-$RECYCLE.BIN/
-
-# Windows Installer files
-*.cab
-*.msi
-*.msm
-*.msp
-
-# Windows shortcuts
-*.lnk
-
-# =========================
-# Operating System Files
-# =========================
-
-# OSX
-# =========================
-
-.DS_Store
-.AppleDouble
-.LSOverride
-
-# Thumbnails
-._*
-
-# Files that might appear in the root of a volume
-.DocumentRevisions-V100
-.fseventsd
-.Spotlight-V100
-.TemporaryItems
-.Trashes
-.VolumeIcon.icns
-
-# Directories potentially created on remote AFP share
-.AppleDB
-.AppleDesktop
-Network Trash Folder
-Temporary Items
-.apdisk
-*.ipch
-AntTracking/x64/Debug/vc140.idb
-*.pdb
-AntTracking/x64/Debug/FlannBasedTracker.obj
-*.obj
-*.tlog
-*.sdf
-*.idb
-*.ilk
-*.log
-*.pch
-AntTracking.VC.opendb
-.vs/AntTracking/v14/.suo
+build/
+externals/
