--- AntTracking/ReadMe.txt
@@ -1,32 +1,0 @@
-﻿========================================================================
-    KONSOLENANWENDUNG: AntTracking-Projektübersicht
-========================================================================
-
-Diese AntTracking-Anwendung wurde vom Anwendungs-Assistenten für Sie erstellt.
-
-Diese Datei bietet eine Übersicht über den Inhalt der einzelnen Dateien, aus
-denen Ihre AntTracking-Anwendung besteht.
-
-
-AntTracking.vcxproj
-    Dies ist die Hauptprojektdatei für VC++-Projekte, die mit dem Anwendungs-Assistenten generiert werden. Sie enthält Informationen über die Version von Visual C++, mit der die Datei generiert wurde, sowie über die Plattformen, Konfigurationen und Projektfunktionen, die im Anwendungs-Assistenten ausgewählt wurden.
-
-AntTracking.vcxproj.filters
-    Dies ist die Filterdatei für VC++-Projekte, die mithilfe eines Anwendungs-Assistenten erstellt werden. Sie enthält Informationen über die Zuordnung zwischen den Dateien im Projekt und den Filtern. Diese Zuordnung wird in der IDE zur Darstellung der Gruppierung von Dateien mit ähnlichen Erweiterungen unter einem bestimmten Knoten verwendet (z. B. sind CPP-Dateien dem Filter "Quelldateien" zugeordnet).
-
-AntTracking.cpp
-    Dies ist die Hauptquelldatei der Anwendung.
-
-/////////////////////////////////////////////////////////////////////////////
-Andere Standarddateien:
-
-StdAfx.h, StdAfx.cpp
-    Mit diesen Dateien werden eine vorkompilierte Headerdatei (PCH) mit dem Namen AntTracking.pch und eine vorkompilierte Typendatei mit dem Namen StdAfx.obj erstellt.
-
-/////////////////////////////////////////////////////////////////////////////
-Weitere Hinweise:
-
-Der Anwendungs-Assistent weist Sie mit "TODO:"-Kommentaren auf Teile des
-Quellcodes hin, die Sie ergänzen oder anpassen sollten.
-
-/////////////////////////////////////////////////////////////////////////////
