--- AntTracking/stdafx.h
@@ -1,15 +1,0 @@
-// stdafx.h: Includedatei für Standardsystem-Includedateien
-// oder häufig verwendete projektspezifische Includedateien,
-// die nur in unregelmäßigen Abständen geändert werden.
-//
-
-#pragma once
-#define _CRT_SECURE_NO_WARNINGS
-#include "targetver.h"
-
-#include <stdio.h>
-#include <tchar.h>
-
-
-
-// TODO: Hier auf zusätzliche Header, die das Programm erfordert, verweisen.
